Bugfix to the RBT, and some portability fixes. test.cpp changed, and so did the hashes. NEW HASHES INCLUDED.
- redblacktreei.cpp: Changed how the masks and bits are handled in the #defines at the top. Now uses the complement operator. This should make for simple portability between x86 and x64.
- odb.cpp: The initial setting of mem_limit seems to overflow the value of a long in 32-bit. Tried explicitly casting to a uint64_t on the right-hand side. Not sure if this solves it.
- test.cpp got another test added. sweeping is now done on the initial ODB object, and on each query result. This changes the file hashes to:
	Batch 1 ........                        6c9bb20e562718fb3c9158e904711a03ca9a897ac355fa801981bfb1b8b27a4a
	Batch 2 (DROP_DUPLICATES) ........      af8333632a28d54ca055f8a82d82c118a3994394203e6672ae0e151c91689f25

+++ odb.cpp
@@ -257,7 +257,7 @@ void ODB::init(DataStore* data, int ident, uint32_t datalen)
     RWLOCK_INIT();
 
     //just to get us started
-    mem_limit=9999999999;
+    mem_limit = (uint64_t)9999999999;
 
     running = 1;
